moved api-test to index
